pyproject.toml
[build-system]
requires = ["poetry-core>=1.4.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "advanced-pytorch"
version = "0.1.0"
description = "Advanced Flower/PyTorch Example"
authors = [
    "The Flower Authors <hello@flower.ai>",
    "Kaushik Amar Das <kaushik.das@iiitg.ac.in>",
]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
hydra-core = "1.3.2"
torch = "1.13.1"
torchvision = "0.14.1"
validators = "0.18.2"
flwr = {extras = ["simulation"], version = "^1.10.0"}
fpzip = "^1.2.4"
cython = "^3.0.11"
scipy = "^1.7.3"